% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteracDiagnosis-package.R
\docType{package}
\name{InteracDiagnosis}
\alias{InteracDiagnosis}
\alias{InteracDiagnosis-package}
\title{InteracDiagnosis}
\description{
A Shiny tool for comparing linear models using coefficient shifts, bootstrap
standard errors, partial plots, and optional cross-validation.
}
\author{
\strong{Maintainer}: Jianyue \email{jianyue.bai@mail.utoronto.ca}

Other contributors:
\itemize{
  \item Selenge [contributor]
}

}
